(chore) move lists-related files into components/Lists subfolders
