+++ frontend/src/components/lists/types/lists.types.ts
@@ -1,0 +1,7 @@
+export type ColorType =
+	| "seagreen"
+	| "teal"
+	| "yellowgreen"
+	| "orange"
+	| "orangered"
+	| "#333";
